+++ plugin/controllers/web.py
@@ -9,7 +9,7 @@
 #                                                                            #
 ##############################################################################
 
-from Components.config import config
+from Components.config import config as comp_config
 from models.info import getInfo, getCurrentTime, getStatusInfo, getFrontendStatus
 from models.services import getCurrentService, getBouquets, getServices, getSubServices, getSatellites, getBouquetEpg, getBouquetNowNextEpg, getServicesNowNextEpg, getSearchEpg, getChannelEpg, getNowNextEpg, getSearchSimilarEpg, getAllServices, getPlayableServices, getPlayableService, getParentalControlList, getEvent, loadEpg, saveEpg
 from models.volume import getVolumeStatus, setVolumeUp, setVolumeDown, setVolumeMute, setVolume
@@ -19,7 +19,7 @@ from models.locations import getLocations, getCurrentLocation, addLocation, remo
 from models.timers import getTimers, addTimer, addTimerByEventId, editTimer, removeTimer, toggleTimerStatus, cleanupTimer, writeTimerList, recordNow, tvbrowser, getSleepTimer, setSleepTimer, getPowerTimer, setPowerTimer, getVPSChannels
 from models.message import sendMessage, getMessageAnswer
 from models.movies import getMovieList, removeMovie, getMovieTags, moveMovie, renameMovie, getAllMovies
-from models.config import getSettings, addCollapsedMenu, removeCollapsedMenu, setZapStream, saveConfig, getZapStream, setShowChPicon, getConfigs, getConfigsSections
+from models.config import getSettings, addCollapsedMenu, removeCollapsedMenu, saveConfig, getConfigs, getConfigsSections
 from models.stream import getStream, getTS, getStreamSubservices, GetSession
 from models.servicelist import reloadServicesLists
 from models.mediaplayer import mediaPlayerAdd, mediaPlayerRemove, mediaPlayerPlay, mediaPlayerCommand, mediaPlayerCurrent, mediaPlayerList, mediaPlayerLoad, mediaPlayerSave, mediaPlayerFindFile
@@ -33,10 +33,10 @@ import re
 
 
 def whoami(request):
-	port = config.OpenWebif.port.value
+	port = comp_config.OpenWebif.port.value
 	proto = 'http'
 	if request.isSecure():
-		port = config.OpenWebif.https_port.value
+		port = comp_config.OpenWebif.https_port.value
 		proto = 'https'
 	ourhost = request.getHeader('host')
 	m = re.match('.+\:(\d+)$', ourhost)
@@ -116,17 +116,17 @@ class WebController(BaseController):
 		success = True
 		oldcheck = False
 		try:
-			if config.usage.check_timeshift.value:
-				oldcheck = config.usage.check_timeshift.value
+			if comp_config.usage.check_timeshift.value:
+				oldcheck = comp_config.usage.check_timeshift.value
 				# don't ask but also don't save
-				config.usage.check_timeshift.value = False
-				config.usage.check_timeshift.save()
+				comp_config.usage.check_timeshift.value = False
+				comp_config.usage.check_timeshift.save()
 			InfoBar.instance.stopTimeshift()
 		except Exception:  # noqa: E722
 			success = False
-		if config.usage.check_timeshift.value:
-			config.usage.check_timeshift.value = oldcheck
-			config.usage.check_timeshift.save()
+		if comp_config.usage.check_timeshift.value:
+			comp_config.usage.check_timeshift.value = oldcheck
+			comp_config.usage.check_timeshift.save()
 		return self.P_tstate(request, success)
 
 	def P_tsstate(self, request, success=True):
@@ -508,7 +508,7 @@ class WebController(BaseController):
 
 		request.setHeader('Content-Type', 'application/x-mpegurl')
 		services = getServices(bRef, False)
-		if config.OpenWebif.auth_for_streaming.value:
+		if comp_config.OpenWebif.auth_for_streaming.value:
 			session = GetSession()
 			if session.GetAuth(request) is not None:
 				auth = ':'.join(session.GetAuth(request)) + "@"
@@ -941,7 +941,7 @@ class WebController(BaseController):
 			HTTP response with headers
 		"""
 		ret = getTimers(self.session)
-		ret["locations"] = config.movielist.videodirs.value
+		ret["locations"] = comp_config.movielist.videodirs.value
 		return ret
 
 	def P_timeradd(self, request):
@@ -1525,8 +1525,8 @@ class WebController(BaseController):
 
 	def P_event(self, request):
 		event = getEvent(request.args["sref"][0], request.args["idev"][0], self.isJson)
-		event['event']['recording_margin_before'] = config.recording.margin_before.value
-		event['event']['recording_margin_after'] = config.recording.margin_after.value
+		event['event']['recording_margin_before'] = comp_config.recording.margin_before.value
+		event['event']['recording_margin_after'] = comp_config.recording.margin_after.value
 		return event
 
 	def P_getcurrent(self, request):
@@ -1656,42 +1656,6 @@ class WebController(BaseController):
 			return res
 		return removeCollapsedMenu(request.args["name"][0])
 
-	def P_zapstream(self, request):
-		"""
-		Request handler for the `zapstream` endpoint.
-
-		.. note::
-
-			Not available in *Enigma2 WebInterface API*.
-
-		Args:
-			request (twisted.web.server.Request): HTTP request object
-		Returns:
-			HTTP response with headers
-		"""
-		res = self.testMandatoryArguments(request, ["checked"])
-		if res:
-			return res
-		return setZapStream(request.args["checked"][0] == "true")
-
-	def P_showchannelpicon(self, request):
-		"""
-		Request handler for the `showchannelpicon` endpoint.
-
-		.. note::
-
-			Not available in *Enigma2 WebInterface API*.
-
-		Args:
-			request (twisted.web.server.Request): HTTP request object
-		Returns:
-			HTTP response with headers
-		"""
-		res = self.testMandatoryArguments(request, ["checked"])
-		if res:
-			return res
-		return setShowChPicon(request.args["checked"][0] == "true")
-
 	def P_streamm3u(self, request):
 		"""
 		Request handler for the `streamm3u` endpoint.
@@ -1715,7 +1679,7 @@ class WebController(BaseController):
 			:query string name: service name
 		"""
 		self.isCustom = True
-		if getZapStream()['zapstream']:
+		if comp_config.OpenWebif.webcache.zapstream.value:
 			if "ref" in request.args:
 				zapService(self.session, request.args["ref"][0], request.args["name"][0], stream=True)
 		return getStream(self.session, request, "stream.m3u")
@@ -2137,37 +2101,21 @@ class WebController(BaseController):
 				})
 		return ret
 
-	def P_settheme(self, request):
-		if "theme" in request.args.keys():
-			theme = request.args["theme"][0]
-			config.OpenWebif.webcache.theme.value = theme
-			config.OpenWebif.webcache.theme.save()
-		return {}
-
 	def P_setmoviesort(self, request):
 		if "nsort" in request.args.keys():
 			nsort = request.args["nsort"][0]
-			config.OpenWebif.webcache.moviesort.value = nsort
-			config.OpenWebif.webcache.moviesort.save()
+			comp_config.OpenWebif.webcache.moviesort.value = nsort
+			comp_config.OpenWebif.webcache.moviesort.save()
 		return {}
 
 	def P_css(self, request):
 		request.setHeader("content-type", "text/css")
 		ret = {}
 		theme = 'original'
-		if config.OpenWebif.webcache.theme.value:
-			theme = config.OpenWebif.webcache.theme.value
+		if comp_config.OpenWebif.webcache.theme.value:
+			theme = comp_config.OpenWebif.webcache.theme.value
 		ret['theme'] = theme
 		return ret
-
-	def P_setmepgmode(self, request):
-		if "mode" in request.args.keys():
-			try:
-				config.OpenWebif.webcache.mepgmode.value = int(request.args["mode"][0])
-				config.OpenWebif.webcache.mepgmode.save()
-			except ValueError:
-				pass
-		return {}
 
 	def P_config(self, request):
 
@@ -2221,6 +2169,35 @@ class WebController(BaseController):
 				# TODO show exception
 				pass
 		return {}
+
+	def P_setwebconfig(self, request):
+		if "responsivedesign" in request.args.keys():
+			val = (request.args["responsivedesign"][0] == 'true')
+			comp_config.OpenWebif.responsive_enabled.value = val
+			comp_config.OpenWebif.responsive_enabled.save()
+		elif "showchannelpicon" in request.args.keys():
+			val = (request.args["showchannelpicon"][0] == 'true')
+			comp_config.OpenWebif.webcache.showchannelpicon.value = val
+			comp_config.OpenWebif.webcache.showchannelpicon.save()
+		elif "zapstream" in request.args.keys():
+			val = (request.args["zapstream"][0] == 'true')
+			comp_config.OpenWebif.webcache.zapstream.value = val
+			comp_config.OpenWebif.webcache.zapstream.save()
+		elif "theme" in request.args.keys():
+			try:
+				comp_config.OpenWebif.webcache.theme.value = request.args["theme"][0]
+				comp_config.OpenWebif.webcache.theme.save()
+			except Exception:
+				pass
+		elif "mepgmode" in request.args.keys():
+			try:
+				comp_config.OpenWebif.webcache.mepgmode.value = int(request.args["mepgmode"][0])
+				comp_config.OpenWebif.webcache.mepgmode.save()
+			except ValueError:
+				pass
+		else:
+			return {"result": False}
+		return {"result": True}
 
 
 class ApiController(WebController):